[Transformations] Move PagedAttentionDecomposition to decompositions/
Relocate the ONNX PagedAttention decomposition pass into the dedicated
decompositions/ directory (per review):
  - src:    transformations/op_conversions/ -> decompositions/
  - header: include/transformations/op_conversions/ ->
            include/openvino/decompositions/ (include path updated to
            openvino/decompositions/paged_attention_decomposition.hpp)
  - test:   tests/op_conversions/ -> tests/decompositions/

Wire the files into DECOMPOSITIONS_SOURCES / DECOMPOSITIONS_HEADERS /
DECOMPOSITIONS_TESTS_SRCS and drop the old op_conversions entries. This also
fixes a pre-existing packaging gap: the header was not enrolled in any header
set in include/sources.cmake, so it now installs like the other public headers.

The pass itself is unchanged (still ov::pass::PagedAttentionDecomposition,
registered in common_optimizations.cpp). Verified: 18 transformation unit
tests and 22 ONNX FE PagedAttention numerical tests (INTERPRETER + IE_CPU)
pass.
